+++ dyatel/dyatel_sel/driver/mobile_driver.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
-import time
 from typing import Union, List
 
 from appium.webdriver.applicationstate import ApplicationState
+from appium.webdriver.common.touch_action import TouchAction
 from appium.webdriver.webdriver import WebDriver as AppiumDriver
+from selenium.webdriver.common.by import By
 
 from dyatel.dyatel_sel.core.core_driver import CoreDriver
 
@@ -17,34 +18,38 @@ class MobileDriver(CoreDriver):
 
         :param driver: appium driver to initialize
         """
-        self.capabilities = driver.capabilities
+        self.caps = driver.capabilities
 
-        self.is_web = self.capabilities.get('browserName', False)
-        self.is_app = self.capabilities.get('app', False)
-        self.is_android = self.capabilities.get('platformName').lower() == 'android'
+        self.is_web = self.caps.get('browserName', False)
+        self.is_app = self.caps.get('app', False)
+        self.is_android = self.caps.get('platformName').lower() == 'android'
 
-        self.is_ios = self.capabilities.get('platformName').lower() == 'ios'
-        self.is_safari_driver = self.capabilities.get('automationName').lower() == 'safari'
-        self.is_xcui_driver = self.capabilities.get('automationName').lower() == 'xcuitest'
+        self.is_ios = self.caps.get('platformName').lower() == 'ios'
+        self.is_safari_driver = self.caps.get('automationName').lower() == 'safari'
+        self.is_xcui_driver = self.caps.get('automationName').lower() == 'xcuitest'
 
-        CoreDriver.mobile = True
         CoreDriver.is_ios = self.is_ios
         CoreDriver.is_android = self.is_android
         CoreDriver.is_safari_driver = self.is_safari_driver
         CoreDriver.is_xcui_driver = self.is_xcui_driver
 
-        self.native_context = 'NATIVE_APP'
-        self.web_context = self.get_web_view_context() if self.is_xcui_driver else 'CHROMIUM'
+        self.native_context_name = 'NATIVE_APP'
+        self.web_context_name = self.get_web_view_context() if self.is_xcui_driver else 'CHROMIUM'
+        self.__is_native_context = None
+        self.__is_web_context = None
+
+        self.top_bar_height = None
+        self.bottom_bar_height = None
 
         if self.is_app:
             if self.is_ios:
-                self.bundle_id = self.capabilities['bundleId']
+                self.bundle_id = self.caps.get('bundleId', 'undefined: bundleId')
             elif self.is_android:
-                self.bundle_id = self.capabilities['appPackage']
+                self.bundle_id = self.caps.get('appPackage', 'undefined: appPackage')
             else:
                 raise Exception('Make sure that correct "platformName" capability specified')
 
-        CoreDriver.__init__(self, driver=driver)
+        super().__init__(driver=driver)
 
     def is_app_installed(self) -> bool:
         """
@@ -109,7 +114,9 @@ class MobileDriver(CoreDriver):
 
         :return: self
         """
-        self.driver.switch_to.context(self.native_context)
+        self.driver.switch_to.context(self.native_context_name)
+        self.__is_native_context = True
+        self.__is_web_context = False
         return self
 
     def switch_to_web(self) -> MobileDriver:
@@ -118,7 +125,9 @@ class MobileDriver(CoreDriver):
 
         :return: self
         """
-        self.driver.switch_to.context(self.web_context)
+        self.driver.switch_to.context(self.web_context_name)
+        self.__is_native_context = False
+        self.__is_web_context = True
         return self
 
     def get_web_view_context(self) -> Union[None, str]:
@@ -139,6 +148,30 @@ class MobileDriver(CoreDriver):
         """
         return self.driver.context
 
+    @property
+    def is_native_context(self) -> bool:
+        """
+        Check is current context is native or not
+
+        :return: bool
+        """
+        if not self.__is_native_context:
+            self.__is_native_context = self.get_current_context() == self.native_context_name
+
+        return self.__is_native_context
+
+    @property
+    def is_web_context(self) -> bool:
+        """
+        Check is current context is web or not
+
+        :return: bool
+        """
+        if not self.__is_web_context:
+            self.__is_web_context = self.get_current_context() == self.web_context_name
+
+        return self.__is_web_context
+
     def get_all_contexts(self) -> List[str]:
         """
         Get the contexts within the current session
@@ -146,3 +179,58 @@ class MobileDriver(CoreDriver):
         :return: list of available contexts
         """
         return self.driver.contexts
+
+    def get_top_bar_height(self) -> int:
+        """
+        iOS only: Get top bar height
+
+        :return: self
+        """
+        if not self.top_bar_height:
+            self.switch_to_native()
+            top_bar = self.driver.find_element(
+                By.XPATH,
+                '//*[contains(@name, "SafariWindow")]/XCUIElementTypeOther[1]/XCUIElementTypeOther/XCUIElementTypeOther'
+            )
+            self.top_bar_height = top_bar.size['height']
+            self.switch_to_web()
+
+        return self.top_bar_height
+
+    def get_bottom_bar_height(self, force: bool = False) -> int:
+        """
+        iOS only: Get bottom bar height
+
+        :param force: get the new value forcly
+        :return: self
+        """
+        if force or not self.top_bar_height:
+            self.switch_to_native()
+
+            bottom_bar = self.driver.find_element(
+                By.XPATH,
+                '//*[@name="CapsuleViewController"]/XCUIElementTypeOther[1]'
+            )
+            self.bottom_bar_height = bottom_bar.size['height']
+            self.switch_to_web()
+
+        return self.top_bar_height
+
+    def click_by_coordinates(self, x: int, y: int, silent: bool = False) -> MobileDriver:
+        """
+        Click by given coordinates
+
+        :param x: tap by given x-axis
+        :param y: tap by given y-axis
+        :param silent: erase log
+        :return: self
+        """
+        if not silent:
+            self.log(f'Tap by given coordinates (x: {x}, y: {y})')
+
+        if self.is_ios:
+            TouchAction(self.driver).tap(x=x, y=y).perform()
+        elif self.is_android:
+            super().click_by_coordinates(x=x, y=y, silent=True)
+
+        return self